URI p-1021 wrong answer %5

+++ URI_midterm/p-1021/src/Main.java
@@ -10,7 +10,7 @@ public class Main {
     public static void main(String[] args) {
         Scanner s = new Scanner(System.in);
 
-        double n = s.nextDouble();
+        float n = s.nextFloat();
 
         System.out.print("NOTAS:\n");
         System.out.printf("%d nota(s) de R$ 100.00\n", (int)(n/100));
